+++ StatsdClient/IStatsd.cs
@@ -28,5 +28,12 @@ namespace StatsdClient
     /// <param name="name"></param>
     /// <param name="value"></param>
     void LogSet(string name, int value);
+    /// <summary>
+    /// Log a raw metric that will not get aggregated on the server.
+    /// </summary>
+    /// <param name="name">The metric name.</param>
+    /// <param name="value">The metric value.</param>
+    /// <param name="epoch">(optional) The epoch timestamp. Leave this blank to have the server assign an epoch for you.</param>
+    void LogRaw(string name, int value, long? epoch = null);
   }
 }